requirements.txt
anthropic>=0.25.0
python-telegram-bot>=20.7
requests>=2.31.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
flask>=3.0.0
python-tapo>=0.3.0
pillow>=10.3.0
